Send what pixels need to be rendered using a "bus".
A bus is just a channel that connects the VM with the new Gfx struct.
The Gfx struct uses the Renderer from sdl2 to render all the pixels
that it receives in the channel. This way we don't need to render
the whole screen in each iteration, only the pixels that have changed.